14274: Move extension to own Swift file.
