+++ src/xcode/ENA/ENA/Source/Extensions/UIView+Utils.swift
@@ -1,0 +1,18 @@
+//
+// 🦠 Corona-Warn-App
+//
+
+import UIKit
+
+extension UIView {
+   func roundCorners(corners: UIRectCorner, radius: CGFloat) {
+		let path = UIBezierPath(
+			roundedRect: bounds,
+			byRoundingCorners: corners,
+			cornerRadii: CGSize(width: radius, height: radius)
+		)
+		let mask = CAShapeLayer()
+		mask.path = path.cgPath
+		layer.mask = mask
+	}
+}
