+++ src/xcode/ENA/ENA/Source/Scenes/Home/Cells/TestRegistration/HomeTestRegistrationTableViewCell.swift
@@ -95,16 +95,3 @@ final class HomeTestRegistrationTableViewCell: UITableViewCell {
 	}
 
 }
-
-extension UIView {
-   func roundCorners(corners: UIRectCorner, radius: CGFloat) {
-		let path = UIBezierPath(
-			roundedRect: bounds,
-			byRoundingCorners: corners,
-			cornerRadii: CGSize(width: radius, height: radius)
-		)
-		let mask = CAShapeLayer()
-		mask.path = path.cgPath
-		layer.mask = mask
-	}
-}
